add more instructions, resolve some nits, remove eslintignore
