+++ .prettierignore
@@ -1,3 +1,2 @@
 build
-tmp
 .github/**/*.md
