+++ src/EXPERIMENTAL__infra-as-code/examples/existingResources.json
@@ -1,0 +1,9 @@
+{
+  "existingResources": {
+    "my-space": {
+      "table": "space",
+      "id": "<INSERT_YOUR_SPACE_ID_HERE>",
+      "spaceId": "<INSERT_YOUR_SPACE_ID_HERE>"
+    }
+  }
+}
